Add command handler and context generator for GPT-friendly context creation
